pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "langchain-postgres>=0.0.15",
    "numpy>=2.3.2",
    "openai>=1.100.2",
    "passlib>=1.7.4",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
